도서 페이지 기능 추가

+++ src/apis/books/index.ts
@@ -1,4 +1,5 @@
 import {
+  FieldValue,
   collection,
   deleteDoc,
   doc,
@@ -8,6 +9,7 @@ import {
   orderBy,
   query,
   runTransaction,
+  serverTimestamp,
   startAfter,
   updateDoc,
   where,
@@ -66,6 +68,20 @@ export async function getBooksBatchBy10({
   }
 }
 
+export async function getBooksCountByUid(uid: string): Promise<number> {
+  try {
+    const booksRef = collection(db, "books");
+    const q = query(booksRef, where("uid", "==", uid));
+    const querySnapshot = await getDocs(q);
+    const count = querySnapshot.size;
+    console.log(`Number of books for uid ${uid}: ${count}`);
+    return count;
+  } catch (error) {
+    console.error("Error fetching books count: ", error);
+    throw error;
+  }
+}
+
 export async function getBookById(bookId: string): Promise<Book> {
   const bookDoc = doc(db, "books", bookId);
   const bookSnapshot = await getDoc(bookDoc);
@@ -105,6 +121,7 @@ interface NewBookData {
   uid: string;
   title: string;
   author: string;
+  category: string;
   genres?: string[];
   content: string;
   areaNo: number;
@@ -139,11 +156,13 @@ export async function createBookWithIndex(
         author: newBookData.author,
         genres: newBookData.genres ?? [],
         content: newBookData.content,
+        category: newBookData.category,
         photoUrl: newBookData.photoUrl,
         areaNo: newBookData.areaNo,
         createdAt: new Date(),
         index: newIndex,
       };
+      console.log(bookData);
 
       transaction.set(newBookRef, bookData);
       return { id: newBookRef.id, index: newIndex };
@@ -163,22 +182,29 @@ export async function updateBookByIndex(bookData: Book): Promise<void> {
   if (querySnapshot.empty) {
     throw new Error(`No book found with index ${bookData.index}`);
   }
-
-  //찾아진 첫 인덱스
-  const bookDoc = querySnapshot.docs[0].ref;
-  //기존 이미지 삭제
-  if (querySnapshot.docs[0].data()?.photoUrl) {
-    deleteFile(querySnapshot.docs[0].data()?.photoUrl);
-  }
-  await updateDoc(bookDoc, {
+  const updateBook: NewBookData & { updatedAt: FieldValue } = {
     uid: bookData.uid,
     title: bookData.title,
     author: bookData.author,
-    genres: bookData.genres,
+    genres: bookData.genres ?? [],
     content: bookData.content,
-    photoUrl: bookData.photoUrl,
-    updatedAt: new Date(),
-  });
+    category: bookData.category,
+    areaNo: bookData.areaNo,
+    updatedAt: serverTimestamp(),
+  };
+  //찾아진 첫 인덱스
+  const bookDoc = querySnapshot.docs[0].ref;
+  //기존 이미지 삭제
+  if (
+    bookData.photoUrl &&
+    querySnapshot.docs[0].data()?.photoUrl !== bookData.photoUrl
+  ) {
+    deleteFile(querySnapshot.docs[0].data()?.photoUrl);
+    updateBook.photoUrl = bookData.photoUrl;
+  }
+  console.log(updateBook);
+
+  await updateDoc(bookDoc, updateBook as Record<string, any>);
 }
 
 export async function deleteBook(bookId: string) {
